Create readme.md
GrainPalette is an AI-based system that classifies rice types using transfer learning. It uses pre-trained CNNs like MobileNetV2 to analyze images of rice grains and identify varieties such as Basmati or Jasmine. The model is trained on augmented datasets, deployed with a user-friendly interface, and helps improve accuracy and efficiency in agricultural quality control.